requirements.txt
flask==2.3.3
pydantic==2.3.0
numpy==1.24.3
torch==2.0.1
transformers==4.31.0
faiss-cpu==1.7.4
langchain==0.0.267
langchain-community==0.0.10
python-dotenv==1.0.0
gunicorn==21.2.0
sentencepiece==0.1.99
protobuf==4.24.2
requests==2.31.0
accelerate==0.21.0
safetensors==0.3.1
